+++ solidity/interfaces/IMailbox.sol
@@ -13,9 +13,9 @@ interface IMailbox {
     function process(bytes calldata _metadata, bytes calldata _message)
         external;
 
-    function count() external view returns (uint256);
+    function count() external view returns (uint32);
 
     function root() external view returns (bytes32);
 
-    function latestCheckpoint() external view returns (bytes32, uint256);
+    function latestCheckpoint() external view returns (bytes32, uint32);
 }
